+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## 0.1
 
 * #1 Bootstrap mechanism
+* #2 Dependency injection bridge